# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * BillEase GST Invoice Generator API
 * OpenAPI spec version: 0.1.0
 */
export interface EmailInvoiceInput {
  to_email: string;
}

export interface EmailResult {
  success?: boolean;
  message?: string;
  error?: string;
}

export interface HealthStatus {
  status: string;
}

export interface Company {
  id: number;
  name: string;
  address: string;
  /** @nullable */
  gstin?: string | null;
  /** @nullable */
  state_name?: string | null;
  /** @nullable */
  state_code?: string | null;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  bank_name?: string | null;
  /** @nullable */
  account_number?: string | null;
  /** @nullable */
  ifsc_code?: string | null;
  /** @nullable */
  branch_name?: string | null;
  /** @nullable */
  logo_url?: string | null;
  /** @nullable */
  declaration_text?: string | null;
  /** @nullable */
  authorized_signatory?: string | null;
  /** @nullable */
  invoice_prefix?: string | null;
  /** @nullable */
  invoice_counter?: number | null;
  /** @nullable */
  watermark_text?: string | null;
  /** @nullable */
  watermark_font?: string | null;
  /** @nullable */
  watermark_opacity?: string | null;
  /** @nullable */
  watermark_color?: string | null;
  /** @nullable */
  watermark_size?: string | null;
  /** @nullable */
  smtp_host?: string | null;
  /** @nullable */
  smtp_port?: string | null;
  /** @nullable */
  smtp_user?: string | null;
  /** @nullable */
  smtp_pass?: string | null;
  /** @nullable */
  smtp_from_name?: string | null;
  /** @nullable */
  smtp_secure?: string | null;
  created_at: string;
  updated_at: string;
}

export interface CompanyInput {
  name: string;
  address: string;
  /** @nullable */
  gstin?: string | null;
  /** @nullable */
  state_name?: string | null;
  /** @nullable */
  state_code?: string | null;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  bank_name?: string | null;
  /** @nullable */
  account_number?: string | null;
  /** @nullable */
  ifsc_code?: string | null;
  /** @nullable */
  branch_name?: string | null;
  /** @nullable */
  logo_url?: string | null;
  /** @nullable */
  declaration_text?: string | null;
  /** @nullable */
  authorized_signatory?: string | null;
  /** @nullable */
  invoice_prefix?: string | null;
  /** @nullable */
  watermark_text?: string | null;
  /** @nullable */
  watermark_font?: string | null;
  /** @nullable */
  watermark_opacity?: string | null;
  /** @nullable */
  watermark_color?: string | null;
  /** @nullable */
  watermark_size?: string | null;
  /** @nullable */
  smtp_host?: string | null;
  /** @nullable */
  smtp_port?: string | null;
  /** @nullable */
  smtp_user?: string | null;
  /** @nullable */
  smtp_pass?: string | null;
  /** @nullable */
  smtp_from_name?: string | null;
  /** @nullable */
  smtp_secure?: string | null;
}

export interface Customer {
  id: number;
  name: string;
  address: string;
  /** @nullable */
  gstin?: string | null;
  /** @nullable */
  state_name?: string | null;
  /** @nullable */
  state_code?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  created_at: string;
}

export interface CustomerInput {
  name: string;
  address: string;
  /** @nullable */
  gstin?: string | null;
  /** @nullable */
  state_name?: string | null;
  /** @nullable */
  state_code?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  code?: string | null;
  /** @nullable */
  hsn_sac?: string | null;
  unit: string;
  rate: number;
  tax_percent: number;
  created_at: string;
}

export interface ProductInput {
  name: string;
  /** @nullable */
  code?: string | null;
  /** @nullable */
  hsn_sac?: string | null;
  unit: string;
  rate: number;
  tax_percent: number;
}

export interface InvoiceItem {
  description: string;
  hsnSac?: string;
  unit: string;
  quantity: number;
  rate: number;
  discountPercent: number;
  taxPercent: number;
  amount: number;
  cgstPercent?: number;
  cgstAmount?: number;
  sgstPercent?: number;
  sgstAmount?: number;
  igstPercent?: number;
  igstAmount?: number;
  total: number;
}

export interface Invoice {
  id: number;
  invoice_number: string;
  invoice_date: string;
  /** @nullable */
  e_way_bill_number?: string | null;
  /** @nullable */
  delivery_note?: string | null;
  /** @nullable */
  payment_terms?: string | null;
  /** @nullable */
  supplier_ref?: string | null;
  /** @nullable */
  other_ref?: string | null;
  /** @nullable */
  buyer_order_number?: string | null;
  /** @nullable */
  order_date?: string | null;
  buyer_name: string;
  buyer_address: string;
  /** @nullable */
  buyer_gstin?: string | null;
  /** @nullable */
  buyer_phone?: string | null;
  /** @nullable */
  buyer_state_name?: string | null;
  /** @nullable */
  buyer_state_code?: string | null;
  /** @nullable */
  dispatch_doc_number?: string | null;
  /** @nullable */
  delivery_note_date?: string | null;
  /** @nullable */
  dispatched_through?: string | null;
  /** @nullable */
  destination?: string | null;
  /** @nullable */
  terms_of_delivery?: string | null;
  subtotal: number;
  discount_amount: number;
  taxable_value: number;
  cgst_total: number;
  sgst_total: number;
  igst_total: number;
  grand_total: number;
  amount_in_words: string;
  payment_mode: string;
  status: string;
  items: InvoiceItem[];
  created_at: string;
  updated_at: string;
}

export interface InvoiceInput {
  invoice_number: string;
  invoice_date: string;
  /** @nullable */
  e_way_bill_number?: string | null;
  /** @nullable */
  delivery_note?: string | null;
  /** @nullable */
  payment_terms?: string | null;
  /** @nullable */
  supplier_ref?: string | null;
  /** @nullable */
  other_ref?: string | null;
  /** @nullable */
  buyer_order_number?: string | null;
  /** @nullable */
  order_date?: string | null;
  buyer_name: string;
  buyer_address: string;
  /** @nullable */
  buyer_gstin?: string | null;
  /** @nullable */
  buyer_phone?: string | null;
  /** @nullable */
  buyer_state_name?: string | null;
  /** @nullable */
  buyer_state_code?: string | null;
  /** @nullable */
  dispatch_doc_number?: string | null;
  /** @nullable */
  delivery_note_date?: string | null;
  /** @nullable */
  dispatched_through?: string | null;
  /** @nullable */
  destination?: string | null;
  /** @nullable */
  terms_of_delivery?: string | null;
  subtotal: number;
  discount_amount: number;
  taxable_value: number;
  cgst_total: number;
  sgst_total: number;
  igst_total: number;
  grand_total: number;
  amount_in_words: string;
  payment_mode: string;
  status: string;
  items: InvoiceItem[];
}

export interface InvoiceListResponse {
  data: Invoice[];
  total: number;
  page: number;
  limit: number;
}

export interface NextInvoiceNumber {
  invoice_number: string;
}

export interface DashboardSummary {
  total_invoices: number;
  monthly_revenue: number;
  tax_collected: number;
  avg_invoice_value: number;
}

export interface MonthlyRevenue {
  month: string;
  revenue: number;
}

export interface DeleteResponse {
  success: boolean;
  message: string;
}

export type ListCustomersParams = {
search?: string;
};

export type ListProductsParams = {
search?: string;
};

export type ListInvoicesParams = {
search?: string;
status?: string;
page?: number;
limit?: number;
};