# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BillEase GST Invoice Generator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices with optional search/filter/pagination
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated invoice list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceListResponse"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create a new invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
...
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get a single invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    put:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
  /invoices/{id}/send-email:
    post:
      operationId: sendInvoiceEmail
      tags: [invoices]
      summary: Send invoice via email using company SMTP settings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailInvoiceInput"
      responses:
        "200":
          description: Email sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailResult"
        "400":
          description: Bad request (SMTP not configured or invalid)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailResult"
  /invoices/{id}/duplicate:
    post:
      operationId: duplicateInvoice
      tags: [invoices]
      summary: Duplicate an invoice with a new number and today's date
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
...
components:
  schemas:
    EmailInvoiceInput:
      type: object
      properties:
        to_email:
          type: string
      required:
        - to_email
    EmailResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        error:
          type: string
...
    InvoiceItem:
      type: object
      properties:
        description:
          type: string
        hsnSac:
          type: string
        unit:
          type: string
        quantity:
          type: number
        rate:
          type: number
        discountPercent:
          type: number
        taxPercent:
          type: number
        amount:
          type: number
        cgstPercent:
          type: number
        cgstAmount:
          type: number
        sgstPercent:
          type: number
        sgstAmount:
          type: number
        igstPercent:
          type: number
        igstAmount:
          type: number
        total:
          type: number
      required:
        - description
        - unit
        - quantity
        - rate
        - discountPercent
        - taxPercent
        - amount
        - total
    Invoice:
      type: object
      properties:
        id:
          type: integer
        invoice_number:
          type: string
        invoice_date:
          type: string
        e_way_bill_number:
          type: ["string", "null"]
        delivery_note:
          type: ["string", "null"]
        payment_terms:
          type: ["string", "null"]
        supplier_ref:
          type: ["string", "null"]
        other_ref:
          type: ["string", "null"]
        buyer_order_number:
          type: ["string", "null"]
        order_date:
          type: ["string", "null"]
        buyer_name:
          type: string
        buyer_address:
          type: string
        buyer_gstin:
          type: ["string", "null"]
        buyer_phone:
          type: ["string", "null"]
        buyer_state_name:
          type: ["string", "null"]
        buyer_state_code:
          type: ["string", "null"]
        dispatch_doc_number:
          type: ["string", "null"]
        delivery_note_date:
          type: ["string", "null"]
        dispatched_through:
          type: ["string", "null"]
        destination:
          type: ["string", "null"]
        terms_of_delivery:
          type: ["string", "null"]
        subtotal:
          type: number
        discount_amount:
          type: number
        taxable_value:
          type: number
        cgst_total:
          type: number
        sgst_total:
          type: number
        igst_total:
          type: number
        grand_total:
          type: number
        amount_in_words:
          type: string
        payment_mode:
          type: string
        status:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - invoice_number
        - invoice_date
        - buyer_name
        - buyer_address
        - subtotal
        - discount_amount
        - taxable_value
        - cgst_total
        - sgst_total
        - igst_total
        - grand_total
        - amount_in_words
        - payment_mode
        - status
        - items
        - created_at
        - updated_at
    InvoiceInput:
      type: object
      properties:
        invoice_number:
          type: string
        invoice_date:
          type: string
        e_way_bill_number:
          type: ["string", "null"]
        delivery_note:
          type: ["string", "null"]
        payment_terms:
          type: ["string", "null"]
        supplier_ref:
          type: ["string", "null"]
        other_ref:
          type: ["string", "null"]
        buyer_order_number:
          type: ["string", "null"]
        order_date:
          type: ["string", "null"]
        buyer_name:
          type: string
        buyer_address:
          type: string
        buyer_gstin:
          type: ["string", "null"]
        buyer_phone:
          type: ["string", "null"]
        buyer_state_name:
          type: ["string", "null"]
        buyer_state_code:
          type: ["string", "null"]
        dispatch_doc_number:
          type: ["string", "null"]
        delivery_note_date:
          type: ["string", "null"]
        dispatched_through:
          type: ["string", "null"]
        destination:
          type: ["string", "null"]
        terms_of_delivery:
          type: ["string", "null"]
        subtotal:
          type: number
        discount_amount:
          type: number
        taxable_value:
          type: number
        cgst_total:
          type: number
        sgst_total:
          type: number
        igst_total:
          type: number
        grand_total:
          type: number
        amount_in_words:
          type: string
        payment_mode:
          type: string
        status:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
      required:
        - invoice_number
        - invoice_date
        - buyer_name
        - buyer_address
        - subtotal
        - discount_amount
        - taxable_value
        - cgst_total
        - sgst_total
        - igst_total
        - grand_total
        - amount_in_words
        - payment_mode
        - status
        - items
    InvoiceListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message