# lib/api-zod/src/generated/types/emailResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * BillEase GST Invoice Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface EmailResult {
  success?: boolean;
  message?: string;
  error?: string;
}